Requify_V2/requirements.txt
accelerate==1.6.0
agno==1.2.6
aiohappyeyeballs==2.6.1
aiohttp==3.11.15
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.9.0
appnope==0.1.4
argcomplete==3.6.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
ascii_colors==0.11.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
boto3==1.38.7
botocore==1.38.7
Byaldi==0.0.7
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cohere==5.15.0
colorama==0.4.6
colpali_engine==0.3.8
comm==0.2.2
configparser==7.2.0
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
datasets==3.5.0
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
docstring_parser==0.16
eval_type_backport==0.2.2
executing==2.2.0
fastavro==1.10.0
fastjsonschema==2.21.1
filelock==3.18.0
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
GPUtil==1.4.0
griffe==1.7.3
groq==0.23.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.1
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipympl==0.9.7
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
lancedb==0.22.0
lightrag-hku==1.3.6
logfire-api==3.14.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mcp==1.7.0
mdurl==0.1.2
mistralai==1.7.0
mistune==3.1.3
ml_dtypes==0.5.1
mpmath==1.3.0
mteb==1.6.35
multidict==6.3.0
multiprocess==0.70.16
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
notebook==7.3.3
notebook_shim==0.2.4
numpy==2.2.4
openai==1.76.2
opentelemetry-api==1.32.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20250327
pdfplumber==0.11.6
peft==0.14.0
pexpect==4.9.0
pillow==11.1.0
pipmaster==0.7.2
platformdirs==4.3.7
prettytable==3.16.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.1
pydantic-ai==0.1.8
pydantic-ai-slim==0.1.8
pydantic-evals==0.1.8
pydantic-graph==0.1.8
pydantic-settings==2.8.1
pydantic_core==2.33.0
Pygments==2.19.1
pyparsing==3.2.3
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytrec_eval-terrier==0.5.6
pytz==2025.2
pyuca==1.2
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.24.0
rsa==4.9.1
s3transfer==0.12.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
Send2Trash==1.8.3
sentence-transformers==4.0.1
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
srsly==2.5.1
sse-starlette==2.3.3
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
typer==0.15.2
types-python-dateutil==2.9.0.20241206
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.2
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
wrapt==1.17.2
XlsxWriter==3.2.3
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
msal
